requirements.txt
flask>=3.0.0
pandas>=2.1.0
openpyxl>=3.1.0
xlrd>=2.0.0
numpy>=1.26.0
gunicorn>=21.0.0
psycopg2-binary>=2.9.0
 